chore(toggleRefinement): rename with correct case
