+++ src/connectors/index.ts
@@ -35,7 +35,7 @@ export { default as connectRatingMenu } from './rating-menu/connectRatingMenu';
 export { default as connectStats } from './stats/connectStats';
 export {
   default as connectToggleRefinement,
-} from './toggleRefinement/connectToggleRefinement';
+} from './toggle-refinement/connectToggleRefinement';
 export { default as connectBreadcrumb } from './breadcrumb/connectBreadcrumb';
 export { default as connectGeoSearch } from './geo-search/connectGeoSearch';
 export { default as connectPoweredBy } from './powered-by/connectPoweredBy';
