+++ src/connectors/toggle-refinement/connectToggleRefinement.js
