+++ src/widgets/toggle-refinement/defaultTemplates.js
